Misc. Optimization and formatting
