+++ src/public/javascripts/dynamicLoading.js
@@ -1,10 +1,12 @@
 /*
 Author: Eric Udlis
-Purpose: Dynamically fill tables with content based off database.JSON
+Purpose: Dynamically fill the dashboard with content based off database.JSON
 */
 const database = require('../../database.json');
 
 console.log(database);
+
+// Dynamic Tables
 
 function createHeaderCol(name, group, units) {
   let header = document.createElement('td'); // Creates the actual DOM element
@@ -78,7 +80,7 @@ module.exports.fillAllTables = function fillAllTables() { // eslint-disable-line
   });
 };
 
-// Loading of Maxs and Mins
+// Dynamic Loading of Maxs and Mins
 
 function getMinCell(sensor) {
   return document.getElementById(`${sensor}Min`);
@@ -113,5 +115,59 @@ module.exports.fillAllBounds = function fillAllBounds(state) { // eslint-disable
   subsystems = Object.keys(database);
   subsystems.forEach((system) => {
     fillTableBounds(system, state);
+  });
+};
+
+
+// Dynamic Dropdowns
+
+
+// code that actually creates the element with the passed in information from fillAllItems
+function createItem(name, group, units) { // eslint-disable-line no-unused-vars
+  let fixedUnits = ` (${units})`; // Adds parenthesis to the units string
+  let fixedName = name.replace(/([a-z\xE0-\xFF])([A-Z\xC0\xDF])/g, '$1 $2') + fixedUnits; // Splits the camel case into two words and adds the units
+  fixedName = fixedName.charAt(0).toUpperCase() + fixedName.slice(1); // Capitalizes first letter
+
+  let header = document.createElement('a'); // Creates the actual DOM element
+  header.href = ''; // Sets the class
+  switch (group) {
+    case 'myDropdown1':
+      header.onclick = function onclick() { // sets the onclick value
+        clone('stoppingDistance');
+        return false;
+      };
+      break;
+    case 'myDropdown2':
+      header.onclick = function onclick() { // sets the onclick value
+        generateLineChartOne(name, fixedName);
+        return false;
+      };
+      break;
+    case 'myDropdown3':
+      header.onclick = function onclick() { // sets the onclick value
+        generateLineChartTwo(name, fixedName);
+        return false;
+      };
+      break;
+    default:
+      break;
+  }
+  header.innerHTML = `${fixedName}`; // Sets value in the box
+  let list = document.getElementById(group);
+  list.appendChild(header);
+}
+
+//
+module.exports.fillAllItems = function fillAllItems() { // eslint-disable-line
+  let subsystems = Object.keys(database); // Create array of each subsystem
+  subsystems.forEach((subsystem) => {
+    let currentSystem = database[subsystem];
+    sensors = Object.keys(currentSystem); // Create an array with all sensors in the subsystem
+
+    sensors.forEach((sensor) => {
+      createItem(`${sensor}`, 'myDropdown1', `${currentSystem[sensor].units}`); // For each sensor create an element
+      createItem(`${sensor}`, 'myDropdown2', `${currentSystem[sensor].units}`); // For each sensor create an element
+      createItem(`${sensor}`, 'myDropdown3', `${currentSystem[sensor].units}`); // For each sensor create an element
+    });
   });
 };
